+++ oklch.js
@@ -4,8 +4,8 @@
  * Cylindrical version of Oklab with cylindrical hue
  *
  * @see {@link https://www.w3.org/TR/css-color-4/#ok-lab}
- * @channel {L} 0 100 Lightness
- * @channel {C} 0 40 Chroma
+ * @channel {L} 0 1 Lightness
+ * @channel {C} 0 0.4 Chroma
  * @channel {H} 0 360 Hue angle in degrees
  * @referred display
  * @dynamic sdr
@@ -18,7 +18,7 @@ var oklch = {
 	name: 'oklch'
 };
 
-// L,C,H -> L,a,b (C 0-40, H 0-360)
+// L,C,H -> L,a,b (C 0-0.4, H 0-360)
 oklch.oklab = (l, c, h) => polarToCart(l, c, h);
 
 // L,a,b -> L,C,H (achromatic hue -> 0)