+++ oklrab.js
@@ -4,9 +4,9 @@
  * Modified version of Oklab using toe mapping
  * Optimized for color picker gamut mapping
  *
- * @channel {L} 0 100 Lightness
- * @channel {a} -40 40 Green-Red axis
- * @channel {b} -40 40 Blue-Yellow axis
+ * @channel {L} 0 1 Lightness
+ * @channel {a} -0.4 0.4 Green-Red axis
+ * @channel {b} -0.4 0.4 Blue-Yellow axis
  * @referred display
  * @dynamic sdr
  */
@@ -21,17 +21,13 @@ var oklrab = {
 };
 
 oklrab.oklab = function (l, a, b) {
-	// OKLrab: L 0-100, a/b ±40
-	// Oklab: L 0-100, a/b ±40
-	// toe/toeInv work on 0-1 scale
-	return [toeInv(l / 100) * 100, a, b];
+	// Native Oklab/OKLrab: L 0-1, a/b ±0.4 (toe/toeInv work on 0-1)
+	return [toeInv(l), a, b];
 };
 
 oklab.oklrab = function (l, a, b) {
-	// Oklab: L 0-100, a/b ±40
-	// OKLrab: L 0-100, a/b ±40
-	// toe/toeInv work on 0-1 scale
-	return [toe(l / 100) * 100, a, b];
+	// Native Oklab/OKLrab: L 0-1, a/b ±0.4 (toe/toeInv work on 0-1)
+	return [toe(l), a, b];
 };
 
 export default oklrab;